pyproject.toml
[tool.poetry]
name = "lotb"
version = "0.0.1"
description = "Lord of the Telegram bots"
authors = ["brokenpip3 <brokenpip3@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
lotb = "lotb.lotb:main"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^21.2"}
feedparser = "^6.0.11"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
types-python-dateutil = "^2.9.0.20241206"

[tool.pytest.ini_options]
pythonpath = ["lotb"]

[tool.ruff]
line-length = 120
indent-width = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
ignore_missing_imports = true
follow_imports_for_stubs = true
disallow_untyped_defs = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
